Adding a sample phi 3.5 config and splitting the unittests into smaller files for pylint
